Create Fake News Detection
Fake news is false information and hoaxes spread through social media and other online media. It is generally done to achieve a political agenda.

In this data science project idea, I will build a model that can accurately detect whether a piece of news is real or fake. I will build a TF-IDF-Vectorizer and use a Passive-Aggressive-Classifier to classify news into “Real” and “Fake”.